映射

+++ java/com/ys/wg/dao/UserDao.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Repository
 public interface UserDao {
 
-    User selectUserById(@Param("userId") Long userId);
+    User selectUserById(@Param("userId") int userId);
 
     User selectUserByPhoneOrEmail(@Param("emailOrPhone") String emailOrPhone, @Param("state") Short state);
