+++ java/com/ys/wg/service/UserService.java
@@ -13,5 +13,5 @@ public interface UserService {
 
     User getUserByPhoneOrEmail(String emailOrPhone, Short state);
 
-    User getUserById(Long userId);
+    User getUserById(int userId);
 }
